[dhcp-server] Disable RX on packet socket
Packet sockets result in a deep copy of packets in the RX path on
all interfaces (when not bound to a specific interface). DHCPv4
server only uses a packet socket to send packets so avoid binding
the packet socket to a protocol which makes the packet socket not
receive any packets which avoids a deep packet copy in the RX
path.

Change-Id: I55c5e0f770f6bdd354bc0d339da842c537820b94
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764787
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>