[devfs] Remove device from Devnode
Remove the ability to get the device from the Devnode. This helps
remove the Device class from devfs.h, which will allow the devfs
implementation to be reused across DFv1 and DFv2.

GetDeviceInfo used to walk the devnodes to get the topological path.
This code has been updated so that it does a string matching on the
GetTopologicalPath function. This should have no functional change.

CompositeTestCase::Topology is checking that a composite device is added
under its primary parent. This is now checked by walking the devnodes
and asserting that the devnode at the end of the path is the same
devnode that lives on the composite device.

The CompositeTestCase also gets the composite device by name instead of
using a magic number based on the way that devices are added.

Bug: 115915
Change-Id: Ia865fedf6766128da5f219bd3b7c5f8ba353a564
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764565
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Ian Fisher <iafisher@google.com>