[ffx][fastboot]: improve error reporting via FIDL
In handle_fastboot_request, now in each match arm when handling a
particular request, do not use the `?` operator when calling
self.interface() or creating a listener, and instead chain all the
Results into a single variable which can be matched on and reported back
to the user.

In the past if the call to `self.interface()?` returned an error (for
any reason) no response would be sent to the caller, causing PEER_CLOSED
errors.

Now those errors are being sent and captured

Change-Id: I3d07ce853451b37f8c1e6fdac8581e145e2968a5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/759023
Reviewed-by: Megan Batty <mgnb@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Colin Nelson <colnnelson@google.com>