[driver_manager] Simplify PrepareProxy
PrepareProxy does nothing if a proxy already exists. Refactor this
function to return early if that's the case.

platform-bus-test has a case where there's a device that exposes
a proxy and allows being bound in multi-composite devices.
That is likely the only driver that depends on this behavior.

Bug: 115915
Change-Id: I866be8b944652b5d3a9a53249c9daa0cb43484f2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/765082
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Ian Fisher <iafisher@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>