[driver_manager] Devices are always created with a controller
Driver Manager's logic is simplified if a device is always created with
its device controller. This ensures that a device will not be connected
to a different device controller later.

This change is happening because devfs wants to get the connection to
the device controller at construction time, which wouldn't be possible
if that connection could change.

Bug: 115915
Change-Id: I4a4093ff023230523699b937c3f830a273f22d46
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/765025
Reviewed-by: Ian Fisher <iafisher@google.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>