[internal_sdk] Include .h from sources in the SDK
This makes sure all headers are made available to the Bazel SDK with
the correct include paths even if they are not explicitly public
headers.

This is a temporary solution to unblock debug_agent build, until we
settle on a long-term solution in the linked-bug.

Effect of this CL is tested in https://fxrev.dev/760743, when several
targets are added into internal_sdk.

Bug: 115708
Change-Id: I8f7a3ae78c768104d48d2976e1becfbbfc1ea968
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/763442
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Jay Zhuang <jayzhuang@google.com>