[mdns] retry queries for additional resources
When InstanceRequestor performs a PTR query, responses sometimes
arrive without the additional resources SRV, TXT and A/AAAA. When
this happens, the InstanceRequestor was querying only once for the
missing records. This CL adds retries so that up to three queries
are sent for missing resources. The first query times out after 1s,
the second after 2s, and the third after 4s.

The ResourceRenewer agent has been repurposed to perform the queries.

Test: new and updated unit tests

Change-Id: I9a172e65ad87612fbe3be987ccd7e39c2f67b1d9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/765725
Reviewed-by: Sundarajan Srinivasan <sundarsrini@google.com>
Commit-Queue: Sundarajan Srinivasan <sundarsrini@google.com>