[Recovery UI] Add screens controller
This CL adds the screen controller and starts the system running with a
keyboard. The keyboard has been chosen because it is already available.

The state has been started off in get network name state which displays
the keyboard, when enter is pressed the state machine switches to state
enter password, when enter is pressed again the state changes again and
screen shows the keyboard again (with a different title) as its
temporary unknown state display.

The above shows that everything is connected up correctly and the
screens switch as they should do.

Screens will be added one CL and a time and the starting state etc will
be changed as necessary.

Bug: b/259274924
Change-Id: Ifa3efbc60af9e535cc81f86395fe37821783d8ac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/763092
Commit-Queue: Kevin Twidle <kpt@google.com>
Reviewed-by: Eric Stone <ecstone@google.com>