[botanist] Handle remote IPv6 addr without scope
When detecting the address of the host from the perspective of the
guest, if the address is an IPv6 address, it is usually a link-local
address with a scope (e.g. 'fe80::1:2:3:4%eth-0' has scope 'eth-0').
Since only link-local IPv6 addresses have a scope, look for the presence
of ':' to determine whether the address is an IPv6 address when
interpolating it as the hostname in a URL.

Bug: 115417
Change-Id: Ia62668bf9d6610e3a35e4a4cd476b9ad58fc7712
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762025
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>