[diagnostics][rust] Include `tracing` target in tags.
This change introduces metatags, which apply tags to events based on
span and event metadata. The only metatag introduced by this change is
`Target`, which applies a tag for the target of the event (typically a
Rust module path).

Metatags are configured in the same way that tags are via
`PublishingOptions`. By default, no metatags are applied, but dependents
can enable metatags via the `init_publishing` function or the
`init_and_detach_ok` macro.

Bug: 102081
Test: Unit tests.
Change-Id: I8a938448ce24e3f55db9d803fc794fb81c53d9ed
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/763724
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Sean Olson <seanolson@google.com>