[ffx] Create gn action for ffx tools
This adds a new gn template for defining ffx subtools with all the
associated metadata necessary, and adapts the ffx-echo tool to use it.

This is quite a bit simpler than the other ffx templates, since it
mostly passes on work to the rustc_binary template, and doesn't try to
be clever about plugins or anything like that. Instead of being magical,
if you want to use the old plugin system in something defined as
ffx_tool, you'll just define the plugin separately and depend on it.

Change-Id: I9475231dbe4855d62cecc5e4c44f723b52f82a51
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/766384
Reviewed-by: Colin Nelson <colnnelson@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>