[devfs] Add DevfsDevice
This is the final step in removing Device from devfs, which
is necessary for reusing devfs across DFv1 and DFv2.

This adds a helper class DevfsDevice, which stores the
topological and protocol nodes that represent a single device.

Device now stores a 'devfs' object that contains the Devnodes.
'self' is renamed to topological_node() and 'link' is renamed to
'protocol_node()'.

Bug: 115915

Change-Id: I31928a3dcf64980a58c731ebf8091c7daa045421
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/765125
Reviewed-by: Tamir Duberstein <tamird@google.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>