[ffx][config] Integrate sdk loading into env ctx
This moves the logic previously in ffx_config's root lib.rs around
loading the sdk into the sdk module, and provides interfaces to it off
the EnvironmentContext.

This allows us to:
- encapsulate the logic for use in future workspace implementations,
- discover the in-tree sdk when running from separate
  subtools when run in standalone mode (and thus can't be informed of
  either ffx' location by a top level ffx invocation, or the sdk
  location by the ffx wrapper script).

I've also re-enabled and fixed up the pbms test that loads an in-tree sdk.

Fixed: 114623
Change-Id: I0868b378b7dfb6fd05fd1ce850d89385d76fffbf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/756125
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>