[ffx] Subtools run standalone can find ffx in sdk
This makes it so that when you run a subtool standalone, and it needs to
spawn a daemon, it will find the ffx host tool in the sdk it was
configured with, if there is one.

Note: this also removes the drop handler from the isolate lib, because
that's not really needed anymore. When the isolate dir is removed, the
daemon will shut it down. This only got brought into this change because
Drop::drop isn't an async-safe context.

Fixed: 114371
Change-Id: I49dc33852d92f3c02dc6ddb065775927339ce2dc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/754848
Reviewed-by: Andrew Davies <awdavies@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>