[usb] Assume USB_INTERFACE_ASSOCIATION bind protocol for IAD
Currently, when the usb composite driver DdkAdds a device corresponding
to an IAD in the configuration, it assumes the USB_INTERFACE bind
protocol. For a given class/subclass/protocol, device bind rules need to
distinguish between binding to an interface, or binding to an
association of interfaces.

This assumes the newly added USB_INTERFACE_ASSOCIATION bind protocol
for devices added from IADs.

Bug: 98137
Change-Id: Iaaf998de68196e5cef6e70615451c045793ba6f6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/753343
Commit-Queue: Sam Hansen <hansens@google.com>
Reviewed-by: Jack Brennen <jbrennen@google.com>