[dhcpv6] Keep values that are missing from Replies
When a Reply message includes an IA but does not include options
(IA Address or IA Prefix) for a previously assigned value (addr/prefix),
keep the previously assigned value. That is, merge the current IA values
with the new IA values.

Note that the only way for the server to instruct a client to
immediately stop using a value is to send a Reply message with
an option for the value (IA Address/Prefix) with a zero valid lifetime.

Bug: 80595
Change-Id: I2483407a9f23afcddd5067794d119b3e96e39c7a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/766142
Reviewed-by: Tony Gong <gongt@google.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>