[dhcpv6] Hide `Instant` type behind type arg
This change allows for the core to use a platform-specific type to hold
time/instant values.

The core does not yet use a platform-specific time/instant value. That
will come in a follow-up change which will allow the core to generate
a platform-specific time/value and then return that back to the
platform.

Bug: 80595
Change-Id: I9171d8f8dc27e7ce9a150da4dabe2bc79604be4d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/766322
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Tony Gong <gongt@google.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>