[debugger] Pretty-print [unordered_]set/map
Removes the existing hardcoded partial pretty-printers for map and set
(these just printed the size).

Add expression code to programmatically pretty-print these objects.

Update the map and set iterator pretty-printers to use the
less-hardcoded scheme. Writes a pretty-printer for
unordered_set::iterator. The unordered_map::iterator pretty-printer
doesn't have the required typedefs to be able to express in C++ outside
of the context of the object itself. I didn't think this case was
important enough to keep a one-off hardcoded case for it so this is still
unsupported.

Increase the max instruction length for an expression system program.
This is required for the longer pretty-printers added.

Change-Id: I56f0e604324df821c8335fb58d23e9b88d02e9ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/763317
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>