[starnix] Enabled BinderLibTest.CheckNoHeaderMappedInUser
Android tests expect binder to use the initial mapped address the first
time it returns to user space, while the previous implementation had a
number of allocation/deallocation that prevented it to happen.

The CL ensures that when nothing is allocated on the shared buffer, the
initial address is used.

Change-Id: I72e809d1ce5ca891fd04148b413207121d2d52f7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/768023
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>