[fdio] Allow safe unbind of local node
When a local node is opened, the lock on the local filesystem is not
held. This allows a local node to be unlinked for the local filesystem
while it is run, which prevent safe cleanup of the local node callback
and context.

To prevent this, this CL adds a lock on the local node that is held
while calling the on_open callback and that is used to clear the
callback when the node is unlinked.

Change-Id: I3377541c02be02dfd4d3d97b424a1f502eb09281
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/763422
Reviewed-by: Tamir Duberstein <tamird@google.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Commit-Queue: Benjamin Lerman <qsr@google.com>