[component_manager] Block destruction on routing task completion
On component destruction, instead of cancelling any in-progress routing
tasks, gate on their completion before destroying the component. This
prevents a scenario where a routing task is cancelled by component
destruction, leading to request loss; the paradigmatic example is a test
component that wrote a log, but the log was lost because test_manager
destroyed the component.

This change attempts to eliminate flakes that emerged with the migration
to event_streams (fxbug.dev/107764).

After landing, we will remove sleeps that were added to deflake the
tests, and monitor them to see if they flake again.

Bug: 76579
Bug: 81980
Bug: 107764
Change-Id: Ibe2bc75b4019024c554e0561760b6e722f741dab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/761894
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Commit-Queue: Gary Bressler <geb@google.com>