[network/tests] Move data race test to errors package
This test exercises a race between a client sending data and the test
closing the interface through which data is being sent. When run with
Fast UDP enabled, it emits error logs, because some amount of messages
have been sent by the client (and are enqueued in the zircon socket
shared with the netstack) but have not yet been read by the netstack and
written to the gVisor endpoint. This write emits an error log once the
device has been closed.

Move the test to the netstack-error-fidl-tests package, which allows
high-severity logs which normally cause tests to fail.

Fixed: 114866
Multiply: netstack-interfaces-error-fidl-test
Change-Id: I0f46dab9dbe75a83d583e9aaf43427d8054af54d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/768063
Reviewed-by: Nick Brown <nickbrow@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>