[dhcpv6] Hold lifetimes with updated time
This change updates the `IaEntry::Assigned` variant to hold a value's
`Lifetimes` along with the time the `Lifetimes` was last updated.

This new state is not yet used in this CL and will be used in a
follow-up.

Bug: 80595
Change-Id: I63ea71ac6f30f19f694971826c17b52a25857585
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/766342
Commit-Queue: Ghanan Gowripalan <ghanan@google.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Tony Gong <gongt@google.com>