[ff][target]: retain port information
The TargetAddr struct now wraps SocketAddr.

This allows for a user to specify a port for a target to connect
from and have that information persisted through the system for
subsequent connections (e.g. if the Fastboot interface via TCP is not
listening on the default port, 5554).

Change-Id: I2dd18ed13fd9a6641c769fa4909e0098d2d0549a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762885
Reviewed-by: Megan Batty <mgnb@google.com>
Commit-Queue: Colin Nelson <colnnelson@google.com>