[brcmfmac] Keep interface until port is removed
The destroy interface sequence allows for the interface to go away for
public use before the netdevice port has been removed. This could cause
calls to the netdevice port to crash because the interface had been
taken away but there were no checks for this. Specifically getting the
MAC address of the interface would crash.

Double fix this by not taking the interface away until the port has been
removed but before the actual interface removal starts. The internal
interface data should be safe up until that point.

Additionally add a check in the get MAC address function to check if the
public interface data is still available and if it's not; return an
all-zero MAC address.

Bug: b/257561072
Test: Multiple calls to stop/start interface
Change-Id: Ie148bf277a2feffa82bccf926e22f22e11e6403a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/767663
Reviewed-by: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>