[dhcpv6] Schedule timer to fire at a specific instant
This prepare for a later change which will only know the instant to
fire a timer. The later CL could always find the duration to the
instant the timer should fire by comparing the instant with the
current time but this is almost guaranteed to introduce a slight
delay in the scheduled time.

Bug: 80595
Change-Id: Ie06bf52a73cb35a471b7b9834d729d06ddd63637
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/766525
Reviewed-by: Tony Gong <gongt@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>