[netstack] Log some endpoint write failures at WARN
Writes to the endpoint can fail with tcpip.ErrNetworkUnreachable if, for
example, a client writing data races with the underlying interface being
removed. Since this is expected to occur in normal usage, log as a
warning instead of an error.

This also allows tests that trigger this condition to pass when they
would otherwise fail due to high-severity logs. See
https://fxbug.dev/114866 for more details.

Bug: 114866
Change-Id: If158a2150b099c6a0e004628bd36d53880e7698e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/768242
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Nick Brown <nickbrow@google.com>