[vdl] Change `fx vdl` warning to an error
As part of the deprecation sequence for `fx vdl`, we introduced a
warning and pause at the beginning of the script to inform users of the
pending removal. This CL escalates the warning to a full error, and
terminates the script instead of continuing on. The execution logic of
the script is unchanged, in case of an emergency situation where the
script is actually still needed; the user can remove the "exit 1" line
and use the script as a stop-gap. Full deletion to follow.

Bug: 113140
Change-Id: I9aa0158687f162bd48f480ee9b0c73254fd919e7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/767027
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Renato Mangini Dias <mangini@google.com>