[nvme] Revert Interrupt Mask Clear (INTMC) register offset fix.
This was "fixed" in fxr/753727, but I now believe this was an error due
to a typo in the NVM Express Base Specification 2.0b. On page 46, the
register's offset is listed as Fh. However, in the detailed description
on page 52, the register's offset is indicated to be 10h. The latter
makes more sense, since the immediately preceding register's (INTMS)
offset is consistently Ch, and it has a length of 4 bytes. Hence, the
offset of Fh must be a typo.

This was discovered while trying to get the NVMe driver working with
legacy interrupts (CL to follow).

Bug: 115873
Change-Id: I184673d9b4fcccc50d290b6a54905a39ae99ad54
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764462
Reviewed-by: Simon Shields <simonshields@google.com>
Commit-Queue: HanBin Yoon <hanbinyoon@google.com>