[sdk] Add golden-file list of released SDK atoms
Adds golden files containing the labels of all SDK atoms released in the
respective SDK. These golden files will be used in CTF and ffx plugins
to verify that their dependencies are released in a public SDK.

Today, CTF depends on the SDK directly to generate the SDK's manifest
files. CTF only uses the gn-label key from these files, so rather than
generating the manifests on every build, the dependency verification
scripts can look at these files.

This also enables ffx to use dependency verification for its plugins.
ffx is in the SDK, so adding a dependency on the SDK causes a cycle in
the build.

Fixed: 87864
Bug: 108159

Change-Id: Iaf8a01f0c7a16890016bd354e4c1f9ecb9aeb1c4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/755666
Reviewed-by: Yuxiao Sun <yuxiaosun@google.com>
Commit-Queue: John Shamoon <johnshamoon@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>