[dhcpv6] Accommodate async UDP delivery
The DHCPv6 client unit tests currently assume that sending a datagram on
a local UDP socket is synchronous, that is, that the remote socket has
received the datagram by the time `send_to` returns. This is no longer a
valid assumption when Fast UDP is enabled.

Instead of assuming the datagram has already been delivered to the
client, explicitly wait for the client to handle the incoming message
after sending it and before moving on with the test.

Fixed: 107856
Multiply: dhcpv6-client-tests
Change-Id: Id6639951f26e6f3061477cd5cf5159f315c3d32f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/767462
Reviewed-by: Tony Gong <gongt@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>