[sherlock][thermal] Add CPU throttle crash reports
Due to a change introduced in tqr/525531 crash reports for CPU
were no longer being generated when the CPU was thermally throttled.
This change adds a thermal_state_handler for the CPU, which will report
metrics to the platform metrics node and generate crash reports if
the system throttles the CPU. Adds a flag so these reports are only
generated for CPU throttling and not audio.

Bug: 111311
Test: fx test power-manager-tests power-manager-integration-test
Change-Id: Icad2ab267b14d762e58191b88c758ff8c8620e29
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/761063
Reviewed-by: Onath Dillinger <claridge@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Kyle Gong <kgong@google.com>