[check-licenses] Generate SPDX files.
New licensing initiative: produce SBOM files for all discrete
software "components". I plan to do this in fuchsia.git with the
help of the github spdx/tools-golang library.

This also updates check-licenses to produce an SPDX json file
by default (in addition to the NOTICE file).

The "Relationship" fields in the resulting SPDX file should not be
trusted for accuracy yet. Building a proper dependency tree is difficult
using the output of "fx gn gen". I will address that in a future CL.

Bug: 112848

Change-Id: Ief7bceb28cb5972b589a644a0f578c05d82aa7f7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/760563
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Yaar Schnitman <yaar@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>