[starnix] Refactor lock in BinderProcess
The goal is to move locks into one or 2 internal states. This CL creates
one state and move a few element in it.

It also refactor the deletion of BinderObject so that no BinderProcess
is lock is owned while it happens, as the BinderObject destructor will
lock its owning process.

Change-Id: Ie92fbbd595dfe94497b609bfcc4444f5e26eb82e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/768722
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>