[fdf][bind] Get aliases in bind rules working
Aliasing bind libraries in bind rules was not working previously because
the alias was not being taken into account when constructing the symbol
table. This change separates out the symbol table keys and values to
account for the possibility that the key is aliased.

Fixed: 95939
Tests: Added unit test for aliasing, and updated the bindlib codegen
integration test to use an alias.

Change-Id: Ifac357d4322f4e918abf08621c9420b195c2aa6f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/766506
Fuchsia-Auto-Submit: Novin Changizi <novinc@google.com>
Reviewed-by: Ian Fisher <iafisher@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>