[assembly] Remove global args from images config
The images config generation is moved to a new template which does not
use any global arguments: generated_images_config_new.

We keep the original template around so that we can assert that the
output of the new template matches that of the old. This is necessary
because the template has been modified quite a bit, and accidentally
changing how images are generated is an impactful mistake.

To simplify the code review, the old template is moved to
generated_images_config_old. This will ensure proper diffing when
generated_images_config_new is renamed to generated_images_config in a
follow-up CL. Otherwise, git gets confused between the two templates.

Bug: 116083
Change-Id: I02fdba7e4b50518191992cce86e016c4d74fc636
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/766784
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>