[assembly] Assembly outputs can be namespaced
Rather than namespacing the outputs by image_name, we allow the user to
specify a specific namespace. This allows the prime images to use
"fuchsia" as the image_name while keeping the outputs in a separate
directory.

We want to use the name "fuchsia" for prime images so that we can re-use
the same images config.

Bug: 116083
Change-Id: Idba991c1ad724d882bce316ac6e144a759d80641
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/768609
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>