[superproject] Roll fuchsia [assembly] Assembly outputs can be namespaced
Rather than namespacing the outputs by image_name, we allow the user to
specify a specific namespace. This allows the prime images to use
"fuchsia" as the image_name while keeping the outputs in a separate
directory.

We want to use the name "fuchsia" for prime images so that we can re-use
the same images config.

Original-Bug: 116083
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/768609
Original-Revision: f27ba8d28053ac44f96dbd29b2b4aec36cb49a2f
GitOrigin-RevId: 7794fb989e7c1a4c9ae72427f7c155d2fedda759
Roller-URL: https://ci.chromium.org/b/8796021799391962337
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I4174e77e2d234f57d6e289ec2d73c3a5d4699113
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/769188