[zither] Add support for syscall return/error types
This introduces the notion of a syscall "return type", which is
currently regarded as the return type of the C vDSO representation. In
the general case, this is any `error` specification (which for our
purposes is limited to zx/status). We also account for the lone case of
wanting a syscall to return a elsewhere declared struct (i.e.,
`zx_string_view_t` in `zx_system_get_version()`).

Bug: 110294
Change-Id: I527023f7e3f43c27dff1fe5bec7eb6f39af46552
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/768843
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>