[kernel][mutex] Optimistic timeslice extension
Instead of disabling pre-emption to ensure there is no gap between
mutex acquisition and setting the timeslice extension, simply set the
timeslice extension initially. In the successful case this simply
avoids the work of installing and removing the preempt disable, and in
the failure case manipulating the timeslice extension is similar in
cost to preempt disable.

Further benefit is had in the case of nested CriticalMutexes as if a
timeslice extension has already been set almost no work, in success or
failure paths, needs to be done.

Run-All-Tests: True
Change-Id: I0d21950912220b431f9f1605548f50dd65c480c9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/763983
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>