[xhci] Simplify enumeration
Separate the first try and retry logic to make it easier to parse. There
is only minor code duplication, and it's much easier to understand the
code.

Remove the error handler and AsyncAutoCall which adds complexity and can
be done in the promise error handler.

Add a test case that checks that the error handler makes the slot
disable command on the correct slot.

Change-Id: Ia3472c3ae0707ff52e61023eff0cbc3fbe09ec9d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/765568
Reviewed-by: Sam Hansen <hansens@google.com>
Commit-Queue: Alex Legg <alexlegg@google.com>