[starnix] Ensure the waiters in BinderProcess is locked with the command queue
It is necessary that notifications and the waiter and the insertion of
elements in the queue are synchronized.

Moreover, if a command is pushed on the process queue while a thread is
waiting for commands, the thread should be woken up.

Change-Id: I1c8d0c6cd81814447dc9b6bdec7ee76ed0d333e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/769382
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>