[fxfs] Add FxfsStorageManager integration tests.
This CL adds a single lifecycle test for FxfsStorageManager along a
normal intended usage. It is not meant to be complete test coverage for
this class; more tests (and tests for unintended usage) will follow in
another CL. We might want to hold off submitting this CL until that one
is approved.

Bug: 114703
Change-Id: If10b1b05bf461202bfb22e4e1a69ae5f657328e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764307
Commit-Queue: James Buckland <jbuckland@google.com>
Reviewed-by: Jody Sankey <jsankey@google.com>