[docs][handle] Improve header comments for zx_handle_t
Make it more clear in the comments around the definition of the
fundamental handle type that the last two bits of a valid handle will
always be set to 1, and that users can take advantage of this fact to
encode information in their handle state, as long as bits are restored
to 1 if the handle value is to be passed to a syscall.

Also, leave a link to the public `zx_handle_t` docs which also
describes this behavior to make it a bit easier for users to find if
they are simply skimming source code.

Change-Id: Id6b658e9b6860e5e234c7cc45a7305c908998fc7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715731
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
API-Review: Adam Barth <abarth@google.com>
Reviewed-by: Marco Vanotti <mvanotti@google.com>