[audio][test] Loosen fidelity tolerance (0.01->1.0)
Further loosen the global fidelity tolerance to reduce the
frequency of flakes. Previous attempt (fxr/769182) loosened
our tolerance from 0.001dB to 0.01dB; unfortunately this did
not sufficiently reduce the frequency of fidelity "flake"
failures in CQ. In the most recent occurrence, a single bin
failed by 0.121dB. This CL increases the tolerance to 1.0 dB,
while we reconsider the value of running these suites in the
"blocking" CQ mix. One option we should consider would be to
move these suites to be "informational" generally, that one
(e.g. audio team) could manually opt-into.

Issue 116506 has been created to track the discussion and
possible change in our approach, and this includes the
re-tightening of this threshold if/when we make larger
changes.

Previous context:
There have been numerous flakes over the past few months where
(typically just 1 or 2) frequency response or sinad measurements are
less than 0.01dB below the expected value. For example, in the flakes
that led to internal CLs 628850 and 628378, sinad was off by at most
0.002dB. This very slightly exceeds the current tolerance of 0.001dB.
By increasing the tolerance to 0.010dB, we should dramatically reduce
the incidence of flakes.

Bug: 116506

Change-Id: I3dc72c0ff51adf498e6d2c3a3e17b346e41ac2c6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/769643
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>