[python_library] Add GN awareness to the files in python libraries
This change switches the 'python_library()' template to use the
'group_with_inputs()' template, which allows it to use a fake
action to document to GN that it has source or input files that
it needs.  This allows it to create a dependency chain back to
those files that is visible to GN, instead of just visible in
depfiles to ninja.

This allows the tests created by the 'python_host_test()'
template to be dirtied when the python library's source files
are dirtied.

Change-Id: If55b0103cacbb3ba271e06c741ef445f4facfdc7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/767705
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>