[ipc] Set default of FUCHSIA_API_LEVEL to __Fuchsia_API_level__
When FUCHSIA_API_LEVEL is not defined but __Fuchsia_API_level__ is, use
the latter instead of failing.

This is needed when the ipc library is used in the Bazel @internal_sdk,
as there is no way to convey the requirement of a custom #define in the
SDK metadata.

Fixed: 115709
Change-Id: Id99ab025d6fa3b0df315dc573f47f381b7e8d49a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/769208
Reviewed-by: Dangyi Liu <dangyi@google.com>
Fuchsia-Auto-Submit: Renato Mangini Dias <mangini@google.com>
Commit-Queue: Renato Mangini Dias <mangini@google.com>
Reviewed-by: Shai Barack <shayba@google.com>