[flatland] Add thread annotations and checks to DisplayCompositor
(Rebase of fxr/750499 on top of new changes)

Add thread annotations, dispatcher checks and restructures locking so
that DisplayCompositor is now (hopefully) race-free.

The locking strategy in this change is currently very course. There is
opportunity to split into multiple locks and lock in a more granular
way, but I'm not sure the current contention for the locks is high
enough to warrant the effort. Filed fxbug.dev/115975 to track it.

Edit: After rebase on fxrev.dev/765423 (making VkRenderer thread-safe)
the locking is a bit less coarse.

Fixed: 112618
Change-Id: Ib072a0b04048afcd50334dabbfde22a721d560a9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764784
Reviewed-by: Emircan Uysaler <emircan@google.com>
Commit-Queue: Mikael Pessa <mikaelpessa@google.com>