[kernel][thread] Add a context switch function
This function is called when a thread is context switched to or away
from. This allows us to delay logic until a context switch occurs.

This is particularly useful for virtualization on x86, where we want to
delay the saving and restoring of MSRs to the point at which a context
switch is required, thereby improving VM exit and entry performance.

Bug: 95763
Change-Id: Ib28f9fb3277884425540234bacdd4bec609a37fa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/767564
Reviewed-by: Corey Tabaka <eieio@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Commit-Queue: Abdulla Kamar <abdulla@google.com>