[check-licenses] Fix issue with autogenerated proto files.
These files are auto-generated when you run the refresh script in the
same directory.

A better solution would be to update the script to insert the proper
copyright information in the generated files, but I will leave that for
a future cleanup task.

Fixed: 116528
Change-Id: I8178f66dc2a3514d5ef7cf09b9f8063e2495c0f0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/769008
Fuchsia-Auto-Submit: Jerry Belton <jcecil@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>
Reviewed-by: Filip Filmar <fmil@google.com>