[fshost] add block watcher pause/resume integration test
Add a test that pauses and resumes the block watcher, confirming that
the block device added when the watcher is paused is ignored and the one
added after it was resumed is processed. This change required a small
refactor to how ramdisks are kept track of in the fixture.

Test: fx test fshost_integration_tests
Multiply: fshost_integration_tests
Change-Id: I2d2ff736aa7dfa3d10bbaa9804204669aa777b0f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/769243
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>
Reviewed-by: Chris Suter <csuter@google.com>